refactor: pagination에서 cursor를 이용하도록 변경
