+++ backend/src/main/java/com/yigongil/backend/ui/doc/StudyApi.java
@@ -22,6 +22,7 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import java.util.List;
+import java.util.Optional;
 import org.springframework.http.ResponseEntity;
 
 @Tag(name = "스터디", description = "스터디 관련 api")
@@ -192,7 +193,7 @@ public interface StudyApi {
     @Operation(summary = "피드 조회")
     ResponseEntity<List<FeedPostResponse>> findFeedPosts(
             @Parameter(description = "조회할 스터디 id", required = true) Long id,
-            @Parameter(description = "페이지", required = true) int page
+            @Parameter(description = "마지막으로 본 피드의 아이디, 첫 요청에서는 필요 없음", allowEmptyValue = true) Optional<Long> oldestFeedPostId
     );
 
     @ApiResponses(
